separated out odict and dataframe
