+++ sciris/utils.py
@@ -3,7 +3,7 @@
 ##############################################################################
 
 import numpy as np
-from collections import OrderedDict as _OD
+
 
 # Handle types and Python 2/3 compatibility
 from six import PY2 as _PY2
@@ -1739,1058 +1739,13 @@ def boxoff(ax=None, removeticks=True, flipticks=True):
 
 
 
-##############################################################################
-### ODICT CLASS
-##############################################################################
-
-class odict(_OD):
-    '''
-    An ordered dictionary, like the OrderedDict class, but supporting list methods like integer 
-    referencing, slicing, and appending.
-    
-    Examples:
-        foo = odict({'ah':3,'boo':4, 'cough':6, 'dill': 8}) # Create odict
-        bar = foo.sorted() # Sort the list
-        assert(bar['boo'] == 4) # Show get item by value
-        assert(bar[1] == 4) # Show get item by index
-        assert((bar[0:2] == [3,4]).all()) # Show get item by slice
-        assert((bar['cough':'dill'] == [6,8]).all()) # Show alternate slice notation
-        assert((bar[[2,1]] == [6,4]).all()) # Show get item by list
-        assert((bar[:] == [3,4,6,8]).all()) # Show slice with everything
-        assert((bar[2:] == [6,8]).all()) # Show slice without end
-        bar[3] = [3,4,5] # Show assignment by item
-        bar[0:2] = ['the', 'power'] # Show assignment by slice -- NOTE, inclusive slice!!
-        bar[[0,2]] = ['cat', 'trip'] # Show assignment by list
-        bar.rename('cough','chill') # Show rename
-        print(bar) # Print results
-    
-    Version: 2018mar29
-    '''
-    
-    def __init__(self, *args, **kwargs):
-        ''' See collections.py '''
-        # Standard OrderedDictionary initialization
-        if len(args)==1 and args[0] is None: args = [] # Remove a None argument
-        _OD.__init__(self, *args, **kwargs) # Standard init
-        return None
-
-    def __slicekey(self, key, slice_end):
-        shift = int(slice_end=='stop')
-        if isinstance(key, _numtype): return key
-        elif type(key) is str: return self.index(key)+shift # +1 since otherwise confusing with names (CK)
-        elif key is None: return (len(self) if shift else 0)
-        else: raise Exception('To use a slice, %s must be either int or str (%s)' % (slice_end, key))
-        return None
-
-
-    def __is_odict_iterable(self, key):
-        ''' Check to see whether the "key" is actually an iterable '''
-        output = type(key)==list or type(key)==type(np.array([])) # Do *not* include dict, since that would be recursive
-        return output
-        
-        
-    def __sanitize_items(self, items):
-        ''' Try to convert the output of a slice to an array, but give up easily and return a list '''
-        try: 
-            output = np.array(items) # Try standard Numpy array...
-            if 'S' in str(output.dtype): # ...but instead of converting to string, convert to object array
-                output = np.array(items, dtype=object)
-        except:
-            output = items # If that fails, just give up and return the list
-        return output
-        
-
-
-    def __getitem__(self, key):
-        ''' Allows getitem to support strings, integers, slices, lists, or arrays '''
-        if isinstance(key, (_stringtype,tuple)):
-            try:
-                output = _OD.__getitem__(self, key)
-                return output
-            except Exception as E: # WARNING, should be KeyError, but this can't print newlines!!!
-                if len(self.keys()): errormsg = '%s\nodict key "%s" not found; available keys are:\n%s' % (repr(E), flexstr(key), '\n'.join([flexstr(k) for k in self.keys()]))
-                else:                errormsg = 'Key "%s" not found since odict is empty'% key
-                raise Exception(errormsg)
-        elif isinstance(key, _numtype): # Convert automatically from float...dangerous?
-            thiskey = self.keys()[int(key)]
-            return _OD.__getitem__(self,thiskey)
-        elif type(key)==slice: # Handle a slice -- complicated
-            try:
-                startind = self.__slicekey(key.start, 'start')
-                stopind = self.__slicekey(key.stop, 'stop')
-                if stopind<startind:
-                    print('Stop index must be >= start index (start=%i, stop=%i)' % (startind, stopind))
-                    raise Exception
-                slicevals = [self.__getitem__(i) for i in range(startind,stopind)]
-                output = self.__sanitize_items(slicevals)
-                return output
-            except:
-                print('Invalid odict slice... returning empty list...')
-                return []
-        elif self.__is_odict_iterable(key): # Iterate over items
-            listvals = [self.__getitem__(item) for item in key]
-            output = self.__sanitize_items(listvals)
-            return output
-        else: # Handle everything else
-            return _OD.__getitem__(self,key)
-        
-        
-    def __setitem__(self, key, value):
-        ''' Allows setitem to support strings, integers, slices, lists, or arrays '''
-        if isinstance(key, (str,tuple)):
-            _OD.__setitem__(self, key, value)
-        elif isinstance(key, _numtype): # Convert automatically from float...dangerous?
-            thiskey = self.keys()[int(key)]
-            _OD.__setitem__(self, thiskey, value)
-        elif type(key)==slice:
-            startind = self.__slicekey(key.start, 'start')
-            stopind = self.__slicekey(key.stop, 'stop')
-            if stopind<startind:
-                errormsg = 'Stop index must be >= start index (start=%i, stop=%i)' % (startind, stopind)
-                raise Exception(errormsg)
-            slicerange = range(startind,stopind)
-            enumerator = enumerate(slicerange)
-            slicelen = len(slicerange)
-            if hasattr(value, '__len__'):
-                if len(value)==slicelen:
-                    for valind,index in enumerator:
-                        self.__setitem__(index, value[valind])  # e.g. odict[:] = arr[:]
-                else:
-                    errormsg = 'Slice "%s" and values "%s" have different lengths! (%i, %i)' % (slicerange, value, slicelen, len(value))
-                    raise Exception(errormsg)
-            else: 
-                for valind,index in enumerator:
-                    self.__setitem__(index, value) # e.g. odict[:] = 4
-        elif self.__is_odict_iterable(key) and hasattr(value, '__len__'): # Iterate over items
-            if len(key)==len(value):
-                for valind,thiskey in enumerate(key): 
-                    self.__setitem__(thiskey, value[valind])
-            else:
-                errormsg = 'Keys "%s" and values "%s" have different lengths! (%i, %i)' % (key, value, len(key), len(value))
-                raise Exception(errormsg)
-        else:
-            _OD.__setitem__(self, key, value)
-        return None
-    
-     
-    def __repr__(self, maxlen=None, showmultilines=True, divider=False, dividerthresh=10, numindents=0, recurselevel=0, sigfigs=None, numformat=None):
-        ''' Print a meaningful representation of the odict '''
-        
-        # Set primitives for display.
-        toolong = ' [...]' # String to display at end of line when maximum value character length is overrun.
-        dividerstr = '*'*40+'\n' # String to use as an inter-item divider.
-        indentstr = '    ' # Create string to use to indent.
-        
-        # Only if we are in the root call, indent by the number of indents.
-        if recurselevel == 0: 
-            theprefix = indentstr * numindents
-        else: # Otherwise (if we are in a recursive call), indent only 1 indent.
-            theprefix = indentstr 
-                            
-        # If the odict is empty, make the string just indicate it's an odict.
-        if len(self.keys())==0:
-            output = 'odict()'
-        else:                   
-            output = '' # Initialize the output to nothing.
-            keystrs = [] # Start with an empty list which we'll save key strings in.
-            valstrs = [] # Start with an empty list which we'll save value strings in.
-            vallinecounts = [] # Start with an empty list which we'll save line counts in.
-            for i in range(len(self)): # Loop over the dictionary values
-                thiskeystr = flexstr(self.keys()[i]) # Grab a str representation of the current key.  
-                thisval = self.values()[i] # Grab the current value.
-                                
-                # If it's another odict, make a call increasing the recurselevel and passing the same parameters we received.
-                if isinstance(thisval, odict):
-                    thisvalstr = flexstr(thisval.__repr__(maxlen=maxlen, showmultilines=showmultilines, divider=divider, 
-                        dividerthresh=dividerthresh, numindents=numindents, recurselevel=recurselevel+1, sigfigs=sigfigs, numformat=numformat))
-                elif isnumber(thisval): # Flexibly print out numbers, since they're largely why we're here
-                    if numformat is not None:
-                        thisvalstr = numformat % thisval
-                    elif sigfigs is not None:
-                        thisvalstr = sigfig(thisval, sigfigs=sigfigs)
-                    else:
-                        thisvalstr = str(thisval) # To avoid numpy's stupid 0.4999999999945
-                else: # Otherwise, do the normal repr() read.
-                    thisvalstr = repr(thisval)
-
-                # Add information to the lists to retrace afterwards.
-                keystrs.append(thiskeystr)
-                valstrs.append(thisvalstr)
-                vallinecounts.append(thisvalstr.count('\n') + 1) # Count the number of lines in the value.
-            maxvallinecounts = max(vallinecounts)   # Grab the maximum count of lines in the dict values.                    
-            
-            maxkeylen = max([len(keystr) for keystr in keystrs])
-            for i in range(len(keystrs)): # Loop over the lists
-                keystr = keystrs[i]
-                valstr = valstrs[i]
-                vallinecount = vallinecounts[i]
-                
-                if (divider or (maxvallinecounts>dividerthresh)) and \
-                    showmultilines and recurselevel==0 and i!=0: # Add a divider line if we should.
-                    newoutput = indent(prefix=theprefix, text=dividerstr, width=80)
-                    if newoutput[-1] == '\n':
-                        newoutput = newoutput[:-1]
-                    output += newoutput        
-                            
-                # Trim the length of the entry if we need to.
-                if not showmultilines:                    
-                    valstr = valstr.replace('\n','\\n') # Replace line breaks with characters
-                
-                # Trim long entries
-                if maxlen and len(valstr) > maxlen: 
-                    valstr = valstr[:maxlen-len(toolong)] + toolong 
-                    
-                # Create the the text to add, apply the indent, and add to the output
-                spacer = ' '*(maxkeylen-len(keystr))
-                if vallinecount == 1 or not showmultilines:
-                    rawoutput = '#%i: "%s":%s %s\n' % (i, keystr, spacer, valstr)
-                else:
-                    rawoutput = '#%i: "%s":%s \n%s\n' % (i, keystr, spacer, valstr)
-                    
-                # Perform the indentation.
-                newoutput = indent(prefix=theprefix, text=rawoutput, width=80)
-                
-                # Strip ot any terminal newline.
-                if newoutput[-1] == '\n':
-                    newoutput = newoutput[:-1] 
-                    
-                # Add the new output to the full output.              
-                output += newoutput                    
-                    
-        # Trim off any terminal '\n'.
-        if output[-1] == '\n':
-            output = output[:-1]
-                
-        # Return the formatted string.
-        return output
-    
-        
-    def _repr_pretty_(self, p, cycle):
-        ''' Function to fix __repr__ in IPython'''
-        print(self.__repr__())
-    
-    
-    def disp(self, maxlen=None, showmultilines=True, divider=False, dividerthresh=10, numindents=0, sigfigs=5, numformat=None):
-        '''
-        Print out flexible representation, short by default.
-        
-        Example:
-            import optima as op
-            import pylab as pl
-            z = op.odict().make(keys=['a','b','c'], vals=(10*pl.rand(3)).tolist())
-            z.disp(sigfigs=3)
-            z.disp(numformat='%0.6f')
-        '''
-        print(self.__repr__(maxlen=maxlen, showmultilines=showmultilines, 
-            divider=divider, dividerthresh=dividerthresh, 
-            numindents=numindents, recurselevel=0, sigfigs=sigfigs, numformat=None))
-    
-    
-    def export(self, doprint=True):
-        ''' Export the odict in a form that is valid Python code '''
-        start = 'odict(['
-        end = '])'
-        output = start
-        
-        for key in self.keys():
-            output += '('+repr(key)
-            output += ', '
-            child = self.get(key)
-            if isinstance(child, odict): output += child.export(doprint=False) # Handle nested odicts -- WARNING, can't doesn't work for e.g. lists of odicts!
-            else:                        output += repr(child)
-            output += '), '
-        
-        output += end
-        if doprint:
-            print(output)
-            return None
-        else:
-            return output
-    
-    
-    def pop(self, key, *args, **kwargs):
-        ''' Allows pop to support strings, integers, slices, lists, or arrays '''
-        if isinstance(key, _stringtype):
-            return _OD.pop(self, key, *args, **kwargs)
-        elif isinstance(key, _numtype): # Convert automatically from float...dangerous?
-            thiskey = self.keys()[int(key)]
-            return _OD.pop(self, thiskey, *args, **kwargs)
-        elif type(key)==slice: # Handle a slice -- complicated
-            try:
-                startind = self.__slicekey(key.start, 'start')
-                stopind = self.__slicekey(key.stop, 'stop')
-                if stopind<startind:
-                    print('Stop index must be >= start index (start=%i, stop=%i)' % (startind, stopind))
-                    raise Exception
-                slicevals = [self.pop(i, *args, **kwargs) for i in range(startind,stopind)] # WARNING, not tested
-                try: return np.array(slicevals) # Try to convert to an array
-                except: return slicevals
-            except:
-                print('Invalid odict slice... returning empty list...')
-                return []
-        elif self.__is_odict_iterable(key): # Iterate over items
-            listvals = [self.pop(item, *args, **kwargs) for item in key]
-            try: return np.array(listvals)
-            except: return listvals
-        else: # Handle string but also everything else
-            try:
-                return _OD.pop(self, key, *args, **kwargs)
-            except: # WARNING, should be KeyError, but this can't print newlines!!!
-                if len(self.keys()): 
-                    errormsg = 'odict key "%s" not found; available keys are:\n%s' % (flexstr(key), 
-                        '\n'.join([flexstr(k) for k in self.keys()]))
-                else: errormsg = 'Key "%s" not found since odict is empty'% key
-                raise Exception(errormsg)
-    
-    
-    def index(self, value):
-        ''' Return the index of a given key '''
-        return self.keys().index(value)
-    
-    
-    def valind(self, value):
-        ''' Return the index of a given value '''
-        return self.items().index(value)
-    
-    
-    def append(self, key=None, value=None):
-        ''' Support an append method, like a list '''
-        needkey = False
-        if value is None: # Assume called with a single argument
-            value = key
-            needkey = True
-        if key is None or needkey:
-            keyname = 'key'+flexstr(len(self))  # Define the key just to be the current index
-        else:
-            keyname = key
-        self.__setitem__(keyname, value)
-        return None
-    
-    
-    def insert(self, pos=None, key=None, value=None):
-        '''
-        Stupid, slow function to do insert -- WARNING, should be able to use approach more like rename...
-        
-        Usage:
-            z = odict()
-            z['foo'] = 1492
-            z.insert(1604)
-            z.insert(0, 'ganges', 1444)
-            z.insert(2, 'midway', 1234)
-        '''
-        
-        # Handle inputs
-        realpos, realkey, realvalue = pos, key, value
-        if key is None and value is None: # Assume it's called like odict.insert(666)
-            realvalue = pos
-            realkey = 'key'+flexstr(len(self))
-            realpos = 0
-        elif value is None: # Assume it's called like odict.insert('devil', 666)
-            realvalue = key
-            realkey = pos
-            realpos = 0
-        if pos is None:
-            realpos = 0
-        if realpos>len(self):
-            errormsg = 'Cannot insert %s at position %i since length of odict is %i ' % (key, pos, len(self))
-            raise Exception(errormsg)
-        
-        # Create a temporary dictionary to hold all of the items after the insertion point
-        tmpdict = odict()
-        origkeys = self.keys()
-        originds = range(len(origkeys))
-        if not len(originds) or realpos==len(originds): # It's empty or in the final position, just append
-            self.__setitem__(realkey, realvalue)
-        else: # Main usage case, it's not empty
-            try: insertind = originds.index(realpos) # Figure out which index we're inseting at
-            except:
-                errormsg = 'Could not insert item at position %i in odict with %i items' % (realpos, len(originds))
-                raise Exception(errormsg)
-            keystopop = origkeys[insertind:] # Pop these keys until we get far enough back
-            for keytopop in keystopop:
-                tmpdict.__setitem__(keytopop, self.pop(keytopop))
-            self.__setitem__(realkey, realvalue) # Insert the new item at the right location
-            for keytopop in keystopop: # Insert popped items back in
-                self.__setitem__(keytopop, tmpdict.pop(keytopop))
-
-        return None
-        
-        
-    def rename(self, oldkey, newkey):
-        ''' Change a key name -- WARNING, very inefficient! '''
-        nkeys = len(self)
-        if isinstance(oldkey, _numtype): 
-            index = oldkey
-            keystr = self.keys()[index]
-        else: # Forge ahead for strings and anything else!
-            index = self.keys().index(oldkey)
-            keystr = oldkey
-        self.__setitem__(newkey, self.pop(keystr))
-        if index<nkeys-1:
-            for i in range(index+1, nkeys):
-                key = self.keys()[index]
-                value = self.pop(key)
-                self.__setitem__(key, value)
-        return None
-    
-    
-    def sort(self, sortby=None, reverse=False, copy=False):
-        '''
-        Create a sorted version of the odict. Sorts by order of sortby, if provided, otherwise alphabetical.
-        If copy is True, then returns a copy (like sorted()).
-        
-        Note that you can also use this to do filtering.
-        
-        Note: very slow, do not use for serious computations!!
-        '''
-        origkeys = self.keys()
-        if sortby is None: allkeys = sorted(origkeys)
-        else:
-            if not isiterable(sortby): raise Exception('Please provide a list to determine the sort order.')
-            if all([isinstance(x,_stringtype) for x in sortby]): # Going to sort by keys
-                allkeys = sortby # Assume the user knows what s/he is doing
-            elif all([isinstance(x,bool) for x in sortby]): # Using Boolean values
-                allkeys = []
-                for i,x in enumerate(sortby):
-                     if x: allkeys.append(origkeys[i])
-            elif all([isinstance(x,_numtype) for x in sortby]): # Going to sort by numbers
-                if not set(sortby)==set(range(len(self))):
-                    errormsg = 'List to sort by "%s" is not compatible with length of odict "%i"' % (sortby, len(self))
-                    raise Exception(errormsg)
-                else: allkeys = [y for (x,y) in sorted(zip(sortby,origkeys))]
-            else: 
-                raise Exception('Cannot figure out how to sort by "%s"' % sortby)
-        tmpdict = odict()
-        if reverse: allkeys.reverse() # If requested, reverse order
-        if copy:
-            for key in allkeys: tmpdict[key] = self[key]
-            return tmpdict
-        else:
-            for key in allkeys: tmpdict.__setitem__(key, self.pop(key))
-            for key in allkeys: self.__setitem__(key, tmpdict.pop(key))
-            return None
-    
-    def sorted(self, sortby=None, reverse=False):
-        ''' Shortcut for making a copy of the sorted odict '''
-        return self.sort(sortby=sortby, copy=True, reverse=reverse)
-
-
-    def reverse(self, copy=False):
-        ''' Reverse the order of an odict '''
-        reversedkeys = self.keys()
-        reversedkeys.reverse()
-        output = self.sort(sortby=reversedkeys, copy=copy)
-        return output
-    
-    
-    def reversed(self):
-        ''' Shortcut for making a copy of the sorted odict '''
-        return self.reverse(copy=True)
-    
-    
-    def make(self, keys=None, vals=None, keys2=None, keys3=None):
-        '''
-        An alternate way of making or adding to an odict. Examples:
-            a = odict().make(5) # Make an odict of length 5, populated with Nones and default key names
-            b = odict().make('foo',34) # Make an odict with a single key 'foo' of value 34
-            c = odict().make(['a','b']) # Make an odict with keys 'a' and 'b'
-            d = odict().make(['a','b'],0) # Make an odict with keys 'a' and 'b', initialized to 0
-            e = odict().make(keys=['a','b'], vals=[1,2]) # Make an odict with 'a':1 and 'b':2
-            f = odict({'a':34, 'b':58}).make(['c','d'],[99,45]) # Add extra keys to an exising odict
-            g = odict().make(keys=['a','b','c'], keys2=['A','B','C'], keys3=['x','y','z'], vals=0) # Make a triply nested odict
-        '''
-        # Handle keys
-        keylist = []
-        if keys is None and vals is None:
-            return None # Nothing to do if nothing supplied
-        if keys is None and vals is not None:
-            keys = len(promotetolist(vals)) # Values are supplied but keys aren't: use default keys
-        if isinstance(keys, _numtype): # It's a single number: pre-generate
-            keylist = ['%i'%i for i in range(keys)] # Generate keylist
-        elif isinstance(keys, _stringtype): # It's a single string
-            keylist = [flexstr(keys)]
-        elif isinstance(keys, list): # It's a list: use directly
-            keylist = keys
-        else:
-            errormsg = 'Could not understand keys "%s": must be number, string, or list' % keys
-            raise Exception(errormsg)
-        nkeys = len(keylist)
-        
-        # Handle values
-        vals = promotetolist(vals)
-        nvals = len(vals)
-        if nvals==0: # Special case: it's an empty list
-            vallist = [dcp(vals) for _ in range(nkeys)]
-        elif nvals==1: # Only a single value: duplicate it
-            vallist = [dcp(vals[0]) for _ in range(nkeys)]
-        elif nvals==nkeys: # Lengths match, can use directly
-            vallist = vals 
-        else:
-            errormsg = 'Must supply either a single value or a list of same length as the keys (%i keys, %i values supplied)' % (nkeys, nvals)
-            raise Exception(errormsg)
-        
-        # Handle nested keys -- warning, would be better to not hard-code this, but does the brain in as it is!
-        if keys2 is not None and keys3 is not None: # Doubly nested
-            self.make(keys=keys, vals=odict().make(keys=keys2, vals=odict().make(keys=keys3, vals=vals)))
-        elif keys2 is not None: # Singly nested
-            self.make(keys=keys, vals=odict().make(keys=keys2, vals=vals))
-        else: # Not nested -- normal case of making an odict
-            for key,val in zip(keylist,vallist): # Update odict
-                self.__setitem__(key, val)
-        
-        return self # A bit weird, but usually would use this return an odict
-    
-    
-    def makefrom(self, source=None, keys=None, keynames=None, *args, **kwargs):
-        '''
-        Create an odict from entries in another dictionary. If keys is None, then
-        use all keys from the current dictionary.
-        
-        Examples:
-            a = 'cat'; b = 'dog'; o = odict().makefrom(source=locals(), keys=['a','b']) # Make use of fact that variables are stored in a dictionary
-            d = {'a':'cat', 'b':'dog'}; o = odict().makefrom(d) # Same as odict(d)
-            l = ['cat', 'monkey', 'dog']; o = odict().makefrom(source=l, keys=[0,2], keynames=['a','b'])
-        '''
-        
-        # Make sure it's iterable
-        if source is not None: # Don't do anything if there's nothing there
-            if not(isiterable(source)): # Make sure it's iterable
-                source = promotetolist(source)
-            elif isinstance(source, _stringtype):
-                source = [source] # Special case -- strings are iterable, but we don't want to
-            
-            if len(source)==0:
-                return self # Nothing to do here
-            else:
-                # Handle cases where keys or keynames are not supplied
-                if keys is None:
-                    if isinstance(source, (list, tuple)):   keys = range(len(source))
-                    elif isinstance(source, dict): keys = source.keys()
-                    else:                          raise Exception('Unable to guess keys for object of type %s' % type(source))
-                keys = promotetolist(keys) # Make sure it's a list
-                if keynames is None: keynames = keys # Use key names
-                
-                # Loop over supplied keys
-                for key,keyname in zip(keys,keynames):
-                    try: 
-                        self.__setitem__(str(keyname), source[key])
-                    except Exception as E: 
-                        raise Exception('Key "%s" not found: %s' % (key, repr(E)))
-                
-        return self # As with make()
-    
-    
-    def map(self, func=None):
-        '''
-        Apply a function to each element of the odict, returning
-        a new odict with the same keys.
-        
-        Example:
-            cat = odict({'a':[1,2], 'b':[3,4]})
-            def myfunc(mylist): return [i**2 for i in mylist]
-            dog = cat.map(myfunc) # Returns odict({'a':[1,4], 'b':[9,16]})
-        '''
-        output = odict()
-        for key in self.keys():
-            output[key] = func(self.__getitem__(key))
-        return output
-    
-    
-    def fromeach(self, ind=None, asdict=True):
-        '''
-        Take a "slice" across all the keys of an odict, applying the same
-        operation to entry. The simplest usage is just to pick an index.
-        However, you can also use it to apply a function to each key.
-        
-        Example:
-            z = odict({'a':array([1,2,3,4]), 'b':array([5,6,7,8])})
-            z.fromeach(2) # Returns array([3,7])
-            z.fromeach(ind=[1,3], asdict=True) # Returns odict({'a':array([2,4]), 'b':array([6,8])})
-        '''
-        output = odict()
-        for key in self.keys():
-            output[key] = self.__getitem__(key)[ind]
-        if asdict: return output # Output as a slimmed-down odict
-        else:      return output[:] # Output as just the entries
-        
-    
-    def toeach(self, ind=None, val=None):
-        '''
-        The inverse of fromeach: partially reset elements within
-        each odict key.
-        
-        Example:
-            z = odict({'a':[1,2,3,4], 'b':[5,6,7,8]})
-            z.toeach(2, [10,20])    # z is now odict({'a':[1,2,10,4], 'b':[5,6,20,8]})
-            z.toeach(ind=3,val=666) #  z is now odict({'a':[1,2,10,666], 'b':[5,6,20,666]})
-        '''
-        nkeys = len(self.keys())
-        if not(isiterable(val)): # Assume it's meant to be populated in each
-            val = [val]*nkeys # Duplicated
-        if len(val)!=nkeys:
-            errormsg = 'To map values onto each key, they must be the same length (%i vs. %i)' % (len(val), nkeys)
-            raise Exception(errormsg)
-        for k,key in self.enumkeys():
-            self.__getitem__(key)[ind] = val[k]
-        return None
-        
-    
-    def enumkeys(self):
-        ''' Shortcut for enumerate(odict.keys()) '''
-        iterator = list(enumerate(self.keys()))
-        return iterator
-    
-    
-    def enumvals(self):
-        ''' Shortcut for enumerate(odict.values()) '''
-        iterator = list(enumerate(self.values()))
-        return iterator
-    
-    
-    def enumitems(self):
-        ''' Returns tuple of 3 things: index, key, value '''
-        iterator = [] # Would be better to not pre-allocate but what can you do...
-        for ind,item in enumerate(self.items()):
-            thistuple = (ind,)+item # Combine into one tuple
-            iterator.append(thistuple)
-        return iterator
-        
-    # Python 3 compatibility
-    if not _PY2:
-        def keys(self):
-            """ Method to get a list of keys as in Python 2. """
-            return list(_OD.keys(self))
-        
-        def values(self):
-            """ Method to get a list of values as in Python 2. """
-            return list(_OD.values(self))
-        
-        def iteritems(self):
-            """ Method to generate an item iterator as in Python 2. """
-            return list(_OD.items(self))
-        
-        
-        
 
         
-
 
 
 
 ##############################################################################
-### DATA FRAME CLASS
-##############################################################################
-
-class dataframe(object):
-    '''
-    A simple data frame, based on simple lists, for simply storing simple data.
-    
-    Example usage:
-        a = dataframe(cols=['x','y'],data=[[1238,2],[384,5],[666,7]]) # Create data frame
-        print a['x'] # Print out a column
-        print a[0] # Print out a row
-        print a[0,'x'] # Print out an element
-        a[0] = [123,6]; print a # Set values for a whole row
-        a['y'] = [8,5,0]; print a # Set values for a whole column
-        a['z'] = [14,14,14]; print a # Add new column
-        a.addcol('z', [14,14,14]); print a # Alternate way to add new column
-        a.rmcol('z'); print a # Remove a column
-        a.pop(1); print a # Remove a row
-        a.append([555,2,14]); print a # Append a new row
-        a.insert(1,[555,2,14]); print a # Insert a new row
-        a.sort(); print a # Sort by the first column
-        a.sort('y'); print a # Sort by the second column
-        a.addrow([555,2,14]); print a # Replace the previous row and sort
-        a.getrow(1) # Return the row starting with value '1'
-        a.rmrow(); print a # Remove last row
-        a.rmrow(1238); print a # Remove the row starting with element '3'
-    
-    Works for both numeric and non-numeric data.
-    
-    Version: 2018mar27
-    '''
-
-    def __init__(self, cols=None, data=None):
-        
-        # Handle columns        
-        if cols is None: cols = list()
-        else:            cols = promotetolist(cols)
-        
-        # Handle data
-        if data is None: 
-            data = np.zeros((0,len(cols)), dtype=object) # Object allows more than just numbers to be stored
-        else:
-            data = np.array(data, dtype=object)
-            if data.ndim != 2:
-                if data.ndim == 1:
-                    if len(cols)==1:
-                        data = np.reshape(data, (len(data),1))
-                    else:
-                        errormsg = 'Dimension of data can only be 1 if there is 1 column, not %s' % len(cols)
-                        raise Exception(errormsg)
-                else:
-                    errormsg = 'Dimension of data must be 1 or 2, not %s' % data.ndim
-                    raise Exception(errormsg)
-            if data.shape[1]==len(cols):
-                pass
-            elif data.shape[0]==len(cols):
-                data = data.transpose()
-            else:
-                errormsg = 'Number of columns (%s) does not match array shape (%s)' % (len(cols), data.shape)
-                raise Exception(errormsg)
-        
-        # Store it        
-        self.cols = cols
-        self.data = data
-        return None
-    
-    def __repr__(self, spacing=2):
-        ''' spacing = space between columns '''
-        if not self.cols: # No keys, give up
-            return ''
-        
-        else: # Go for it
-            outputlist = dict()
-            outputformats = dict()
-            
-            # Gather data
-            nrows = self.nrows()
-            for c,col in enumerate(self.cols):
-                outputlist[col] = list()
-                maxlen = len(col) # Start with length of column name
-                if nrows:
-                    for val in self.data[:,c]:
-                        output = flexstr(val)
-                        maxlen = max(maxlen, len(output))
-                        outputlist[col].append(output)
-                outputformats[col] = '%'+'%i'%(maxlen+spacing)+'s'
-            
-            indformat = '%%%is' % (np.floor(np.log10(nrows))+1) # Choose the right number of digits to print
-            
-            # Assemble output
-            output = indformat % '' # Empty column for index
-            for col in self.cols: # Print out header
-                output += outputformats[col] % col
-            output += '\n'
-            
-            for ind in range(nrows): # Loop over rows to print out
-                output += indformat % flexstr(ind)
-                for col in self.cols: # Print out data
-                    output += outputformats[col] % outputlist[col][ind]
-                output += '\n'
-            
-            return output
-    
-    def _val2row(self, value=None):
-        ''' Convert a list, array, or dictionary to the right format for appending to a dataframe '''
-        if isinstance(value, dict):
-            output = np.zeros(self.ncols(), dtype=object)
-            for c,col in enumerate(self.cols):
-                try: 
-                    output[c] = value[col]
-                except: 
-                    errormsg = 'Entry for column %s not found; keys you supplied are: %s' % (col, value.keys())
-                    raise Exception(errormsg)
-            output = np.array(output, dtype=object)
-        elif value is None:
-            output = np.empty(self.ncols(),dtype=object)
-        else: # Not sure what it is, just make it an array
-            if len(value)==self.ncols():
-                output = np.array(value, dtype=object)
-            else:
-                errormsg = 'Row has wrong length (%s supplied, %s expected)' % (len(value), self.ncols())
-        return output
-    
-    def _sanitizecol(self, col):
-        ''' Take None or a string and return the index of the column '''
-        if col is None: output = 0 # If not supplied, assume first column is control
-        elif isinstance(col, _stringtype): output = self.cols.index(col) # Convert to index
-        else: output = col
-        return output
-    
-    def __getitem__(self, key=None):
-        ''' Simple method for returning; see self.get() for a version based on col and row '''
-        if isinstance(key, _stringtype):
-            colindex = self.cols.index(key)
-            output = self.data[:,colindex]
-        elif isinstance(key, _numtype):
-            rowindex = int(key)
-            output = self.data[rowindex,:]
-        elif isinstance(key, tuple):
-            colindex = self.cols.index(key[0])
-            rowindex = int(key[1])
-            output = self.data[rowindex,colindex]
-        elif isinstance(key, slice):
-            rowslice = key
-            slicedata = self.data[rowslice,:]
-            output = dataframe(cols=self.cols, data=slicedata)
-        else:
-            raise Exception('Unrecognized dataframe key "%s"' % key)
-        return output
-        
-    def __setitem__(self, key, value=None):
-        if value is None:
-            value = np.zeros(self.nrows(), dtype=object)
-        if isinstance(key, _stringtype): # Add column
-            if len(value) != self.nrows(): 
-                errormsg = 'Vector has incorrect length (%i vs. %i)' % (len(value), self.nrows())
-                raise Exception(errormsg)
-            else:
-                val_arr = np.reshape(value, (len(value),1))
-            try:
-                colindex = self.cols.index(key)
-                self.data[:,colindex] = val_arr
-            except:
-                self.cols.append(key)
-                colindex = self.cols.index(key)
-                self.data = np.hstack((self.data, np.array(val_arr, dtype=object)))
-        elif isinstance(key, _numtype):
-            value = self._val2row(value) # Make sure it's in the correct format
-            if len(value) != self.ncols(): 
-                errormsg = 'Vector has incorrect length (%i vs. %i)' % (len(value), self.ncols())
-                raise Exception(errormsg)
-            rowindex = int(key)
-            try:
-                self.data[rowindex,:] = value
-            except:
-                self.data = np.vstack((self.data, np.array(value, dtype=object)))
-        elif isinstance(key, tuple):
-            try:
-                colindex = self.cols.index(key[0])
-                rowindex = int(key[1])
-                self.data[rowindex,colindex] = value
-            except:
-                errormsg = 'Could not insert element (%s,%s) in dataframe of shape %' % (key[0], key[1], self.data.shape)
-                raise Exception(errormsg)
-        return None
-    
-    def get(self, cols=None, rows=None):
-        '''
-        More complicated way of getting data from a dataframe; example:
-        df = dataframe(cols=['x','y','z'],data=[[1238,2,-1],[384,5,-2],[666,7,-3]]) # Create data frame
-        df.get(cols=['x','z'], rows=[0,2])
-        '''
-        if cols is None:
-            colindices = Ellipsis
-        else:
-            colindices = []
-            for col in promotetolist(cols):
-                colindices.append(self._sanitizecol(col))
-        if rows is None:
-            rowindices = Ellipsis
-        else:
-            rowindices = rows
-        
-        output = self.data[:,colindices][rowindices,:] # Split up so can handle non-consecutive entries in either
-        if output.size == 1: output = output[0] # If it's a single element, return the value rather than the array
-        return output
-   
-    def pop(self, key, returnval=True):
-        ''' Remove a row from the data frame '''
-        rowindex = int(key)
-        thisrow = self.data[rowindex,:]
-        self.data = np.vstack((self.data[:rowindex,:], self.data[rowindex+1:,:]))
-        if returnval: return thisrow
-        else:         return None
-    
-    def append(self, value):
-        ''' Add a row to the end of the data frame '''
-        value = self._val2row(value) # Make sure it's in the correct format
-        self.data = np.vstack((self.data, np.array(value, dtype=object)))
-        return None
-    
-    def ncols(self):
-        ''' Get the number of columns in the data frame '''
-        ncols = len(self.cols)
-        ncols2 = self.data.shape[1]
-        if ncols != ncols2:
-            errormsg = 'Dataframe corrupted: %s columns specified but %s in data' % (ncols, ncols2)
-            raise Exception(errormsg)
-        return ncols
-
-    def nrows(self):
-        ''' Get the number of rows in the data frame '''
-        try:    return self.data.shape[0]
-        except: return 0 # If it didn't work, probably because it's empty
-    
-    def addcol(self, key=None, value=None):
-        ''' Add a new column to the data frame -- for consistency only '''
-        self.__setitem__(key, value)
-    
-    def rmcol(self, key):
-        ''' Remove a column from the data frame '''
-        colindex = self.cols.index(key)
-        self.cols.pop(colindex) # Remove from list of columns
-        self.data = np.hstack((self.data[:,:colindex], self.data[:,colindex+1:])) # Remove from data
-        return None
-    
-    def addrow(self, value=None, overwrite=True, col=None, reverse=False):
-        ''' Like append, but removes duplicates in the first column and resorts '''
-        value = self._val2row(value) # Make sure it's in the correct format
-        col   = self._sanitizecol(col)
-        index = self._rowindex(key=value[col], col=col, die=False) # Return None if not found
-        if index is None or not overwrite: self.append(value)
-        else: self.data[index,:] = value # If it exists already, just replace it
-        self.sort(col=col, reverse=reverse) # Sort
-        return None
-    
-    def _rowindex(self, key=None, col=None, die=False):
-        ''' Get the sanitized row index for a given key and column '''
-        col = self._sanitizecol(col)
-        coldata = self.data[:,col] # Get data for this column
-        if key is None: key = coldata[-1] # If not supplied, pick the last element
-        try:    index = coldata.tolist().index(key) # Try to find duplicates
-        except: 
-            if die: raise Exception('Item %s not found; choices are: %s' % (key, coldata))
-            else:   return None
-        return index
-        
-    def rmrow(self, key=None, col=None, returnval=False, die=True):
-        ''' Like pop, but removes by matching the first column instead of the index -- WARNING, messy '''
-        index = self._rowindex(key=key, col=col, die=die)
-        if index is not None: self.pop(index)
-        return None
-    
-    def _diffindices(self, indices=None):
-        ''' For a given set of indices, get the inverse, in set-speak '''
-        if indices is None: indices = []
-        ind_set = set(np.array(indices))
-        all_set = set(np.arange(self.nrows()))
-        diff_set = np.array(list(all_set - ind_set))
-        return diff_set
-    
-    def rmrows(self, indices=None):
-        ''' Remove rows by index -- WARNING, messy '''
-        keep_set = self._diffindices(indices)
-        self.data = self.data[keep_set,:]
-        return None
-    
-    def replace(self, col=None, old=None, new=None):
-        ''' Replace all of one value in a column with a new value '''
-        col = self._sanitizecol(col)
-        coldata = self.data[:,col] # Get data for this column
-        inds = findinds(coldata==old)
-        self.data[inds,col] = new
-        return None
-        
-    
-    def _todict(self, row):
-        ''' Return row as a dict rather than as an array '''
-        if len(row)!=len(self.cols): 
-            errormsg = 'Length mismatch between "%s" and "%s"' % (row, self.cols)
-            raise Exception(errormsg)
-        rowdict = dict(zip(self.cols, row))
-        return rowdict
-    
-    def findrow(self, key=None, col=None, default=None, closest=False, die=False, asdict=False):
-        '''
-        Return a row by searching for a matching value.
-        
-        Arguments:
-            key = the value to look for
-            col = the column to look for this value in
-            default = the value to return if key is not found (overrides die)
-            closest = whether or not to return the closest row (overrides default and die)
-            die = whether to raise an exception if the value is not found
-            asdict = whether to return results as dict rather than list
-        
-        Example:
-            df = dataframe(cols=['year','val'],data=[[2016,0.3],[2017,0.5]])
-            df.findrow(2016) # returns array([2016, 0.3], dtype=object)
-            df.findrow(2013) # returns None, or exception if die is True
-            df.findrow(2013, closest=True) # returns array([2016, 0.3], dtype=object)
-            df.findrow(2016, asdict=True) # returns {'year':2016, 'val':0.3}
-        '''
-        if not closest: # Usual case, get 
-            index = self._rowindex(key=key, col=col, die=(die and default is None))
-        else:
-            col = self._sanitizecol(col)
-            coldata = self.data[:,col] # Get data for this column
-            index = np.argmin(abs(coldata-key)) # Find the closest match to the key
-        if index is not None:
-            thisrow = self.data[index,:]
-            if asdict:
-                thisrow = self._todict(thisrow)
-        else:
-            thisrow = default # If not found, return as default
-        return thisrow
-    
-    def rowindex(self, key=None, col=None):
-        ''' Return the indices of all rows matching the given key in a given column. '''
-        col = self._sanitizecol(col)
-        coldata = self.data[:,col] # Get data for this column
-        indices = findinds(coldata==key)
-        return indices
-        
-    def _filterrows(self, key=None, col=None, keep=True, verbose=False):
-        ''' Filter rows and either keep the ones matching, or discard them '''
-        indices = self.rowindex(key=key, col=col)
-        if keep: indices = self._diffindices(indices)
-        self.rmrows(indices=indices)
-        if verbose: print('Dataframe filtering: %s rows removed based on key="%s", column="%s"' % (len(indices), key, col))
-        return None
-    
-    def filter_in(self, key=None, col=None, verbose=False):
-        self._filterrows(key=key, col=col, keep=True, verbose=verbose)
-        return None
-    
-    def filter_out(self, key=None, col=None, verbose=False):
-        self._filterrows(key=key, col=col, keep=False, verbose=verbose)
-        return None
-        
-    def insert(self, row=0, value=None):
-        ''' Insert a row at the specified location '''
-        rowindex = int(row)
-        value = self._val2row(value) # Make sure it's in the correct format
-        self.data = np.vstack((self.data[:rowindex,:], value, self.data[rowindex:,:]))
-        return None
-    
-    def sort(self, col=None, reverse=False):
-        ''' Sort the data frame by the specified column '''
-        col = self._sanitizecol(col)
-        sortorder = np.argsort(self.data[:,col])
-        if reverse: sortorder = np.array(list(reversed(sortorder)))
-        self.data = self.data[sortorder,:]
-        return None
-        
-    def jsonify(self, cols=None, rows=None, header=None):
-        ''' Export the dataframe to a JSON-compatible format '''
-        
-        # Handle input arguments
-        if cols   is None: cols   = self.cols # Use all columns by default
-        if rows   is None: rows   = list(range(self.nrows())) # Use all rows by default
-        if header is None: header = True # Include headers
-        
-        # Handle output
-        output = []
-        if header: output.append(cols)
-        for r in rows:
-            thisrow = []
-            for col in cols:
-                datum = self.get(cols=col,rows=r)
-                thisrow.append(datum)
-            output.append(thisrow)
-        return output
-    
-    def export(self, filename=None, sheetname=None, close=True):
-        from sciris.fileio import export_file
-        for_export = dcp(self.data)
-        for_export = np.vstack((self.cols, self.data))
-        export_file(filename=filename, data=for_export, sheetname=sheetname, close=close)
-        return
-        
-
-
-
-
-
-
-
-
-
-
-##############################################################################
-### OTHER CLASSES
+### LINKS
 ##############################################################################
